--- projects/01-rubygems-cli/solution/errors/gem_not_found_error.rb
@@ -1,2 +1,0 @@
-class GemNotFoundError < StandardError
-end
