--- projects/01-rubygems-cli/solution/errors/standard_api_error.rb
@@ -1,2 +1,0 @@
-class StandardAPIError < StandardError
-end
