--- projects/01-rubygems-cli/solution/lib/command_results/command_result.rb
@@ -1,10 +1,0 @@
-class CommandResult
-  attr_accessor :exit_code
-
-  def initialize(exit_code)
-    @exit_code = exit_code
-  end
-
-  def output
-  end
-end
