--- projects/01-rubygems-cli/solution/lib/command_results/search_command_result.rb
@@ -1,19 +1,0 @@
-require "./lib/command_results/command_result"
-
-class SearchCommandResult < CommandResult
-  attr_accessor :gems
-
-  def initialize(gems, exit_code = 0)
-    super(exit_code)
-    @gems = gems
-  end
-
-  def output
-    result = ""
-    gems.each_with_index do |gem, i|
-      result += "#{i + 1}. #{gem.name}\n"
-    end
-    
-    result
-  end
-end
