+++ projects/01-rubygems-cli/solution/lib/help/command.rb
@@ -1,0 +1,11 @@
+require "./lib/help/result"
+
+module Help
+  class Command
+    class << self
+      def execute
+        Help::Result.new
+      end
+    end
+  end
+end
