+++ projects/01-rubygems-cli/solution/lib/ruby_gems_api/gem_not_found_error.rb
@@ -1,0 +1,4 @@
+module RubyGemsApi
+  class GemNotFoundError < StandardError
+  end
+end
