+++ projects/01-rubygems-cli/solution/lib/ruby_gems_api/standard_api_error.rb
@@ -1,0 +1,4 @@
+module RubyGemsApi
+  class StandardAPIError < StandardError
+  end
+end
